Divide accounting into multiple files
Also rename accesstoken
